Implemented Client for Socket connections, might need improvement
Implemented a Client that starts a new Socket in a new Thread and sends a message to a Server. Basic principle works. Also tried to implement a monitor to tell user about connection state with Toast (personal idea), so far doesn't work as intended, e.g. locking other Client Threads or crash.